Refactor: Renamed all components with prefix "app"
