+++ src/components/AppHeader/app-header.css
@@ -1,0 +1,19 @@
+.app-header {
+  display: flex;
+  width: 100%;
+  height: 40px;
+  background: #ffb763;
+  border-bottom: 2px solid #cb9745;
+}
+
+.app-header__logo {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  padding-left: 10px;
+  padding-right: 10px;
+  font-family: Arial;
+  font-size: 25px;
+  color: #9c7c30;
+}
