+++ src/components/AppPlayground/app-playground.css
@@ -1,3 +1,3 @@
-#playground {
+#app-playground {
   height: 100%;
 }
