+++ src/components/AppSideBar/app-side-bar.css
@@ -1,4 +1,4 @@
-.tool-bar {
+.app-side-bar {
   display: flex;
   justify-content: center;
   width: 100%;
@@ -6,19 +6,12 @@
   border-left: 2px solid #d2d2d2;
 }
 
-.tool-bar__main-menu {
+.app-tool-bar__main-menu {
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  width: 100%;
   height: 30%;
-}
-
-.tool-bar__row {
-  width: 100%;
-  margin-top: 10px;
-}
-
-.tool-bar__action-button {
-  width: 100%;
+  background-color: dimgrey;
 }
