--- src/components/Header/app-header.css
@@ -1,6 +1,0 @@
-.header {
-  width: 100%;
-  height: 40px;
-  background: #ffb763;
-  border-bottom: 2px solid #cb9745;
-}
